Reorganize the Monaco Editor facade
Split the code into multiple files which resides in packages matching
the modules of the API.
