+++ metadoc-js/src/main/scala/monaco/languages/package.scala
@@ -1,0 +1,8 @@
+package monaco
+
+import scala.scalajs.js
+import scala.scalajs.js.|
+
+package object languages {
+  type Definition = Location | js.Array[Location]
+}
